Better organization and modularization - Implement the compute_pointpca2 function on lib.rs - Make the main.rs example use its own package and the compute_pointpca2 function - Separate auxiliary functions from the features.rs module - The eigenvectors.rs module is now focused on anything related to eigenvectors/PCA
