+++ src/main.rs
@@ -1,44 +1,20 @@
-extern crate kd_tree;
-extern crate libm;
-extern crate nalgebra as na;
-extern crate ordered_float;
-extern crate ply_rs;
-extern crate rayon;
-
-mod features;
-mod knn_search;
-mod ply_manager;
-mod pooling;
-mod predictors;
-mod preprocessing;
-mod spatial_metrics;
-mod utils;
+use pointpca2_rust;
+use pointpca2_rust::ply_manager;
 
 fn main() {
     let search_size = 81;
+    let verbose = true;
     println!("Reading ply");
     let (points_a, colors_a) = ply_manager::read_ply_as_matrix("<path-to-reference>");
     let (points_b, colors_b) = ply_manager::read_ply_as_matrix("<path-to-test>");
-    println!("Preprocessing");
-    let (points_a, colors_a) = preprocessing::preprocess_point_cloud(&points_a, &colors_a);
-    let (points_b, colors_b) = preprocessing::preprocess_point_cloud(&points_b, &colors_b);
-    println!("Performing knn search");
-    let knn_indices_a = knn_search::knn_search(&points_a, &points_a, search_size);
-    let knn_indices_b = knn_search::knn_search(&points_b, &points_a, search_size);
-    println!("Computing local features");
-    let local_features = features::compute_features(
+    let pooled_predictors = pointpca2_rust::compute_pointpca2(
         points_a,
         colors_a,
         points_b,
         colors_b,
-        knn_indices_a,
-        knn_indices_b,
         search_size,
+        verbose,
     );
-    println!("Computing predictors");
-    let predictors_result = predictors::compute_predictors(local_features);
-    println!("Pooling predictors");
-    let pooled_predictors = pooling::mean_pooling(predictors_result);
     println!("Predictors:");
     for col in pooled_predictors.iter() {
         print!("{:.4}  ", *col);
